Implemented Inventory Limit
Added a variable that indicates if the inventory is full.
Added if statement that changes previous variable if there are 3 items on the player.
Added if statement that checked if the inventory is and if so, runs a coroutine.
Added a coroutine that displays text indicating to the player the inventory is full and then, after a seconds, taking that text away.